requirements.txt
fastapi==0.136.0
motor==3.7.1
passlib==1.7.4
pydantic==2.13.3
python-dotenv==1.2.2
redis==7.4.0
SQLAlchemy==2.0.49
